--- app/src/main/java/com/techcourse/exception/HandlerFieldException.java
@@ -1,8 +1,0 @@
-package com.techcourse.exception;
-
-public class HandlerFieldException extends RuntimeException {
-
-    public HandlerFieldException() {
-        super("Handler needs Field name 'controllers'");
-    }
-}
