+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/exception/NoSuchHandlerFoundException.java
@@ -1,4 +1,4 @@
-package com.techcourse.exception;
+package webmvc.org.springframework.web.servlet.exception;
 
 public class NoSuchHandlerFoundException extends RuntimeException {
