+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/DispatcherServlet.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -7,6 +7,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webmvc.org.springframework.web.servlet.ModelAndView;
+import webmvc.org.springframework.web.servlet.mvc.adaptor.HandlerAdaptors;
+import webmvc.org.springframework.web.servlet.mvc.adaptor.HandlerMappings;
 
 public class DispatcherServlet extends HttpServlet {
