+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adaptor/AnnotationHandlerAdaptor.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.adaptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
