+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adaptor/HandlerAdaptor.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package webmvc.org.springframework.web.servlet.mvc.adaptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
